VSR: Change cluster id from u32 to u128
Rationale (via https://github.com/tigerbeetle/tigerbeetle/pull/1295#discussion_r1389061252):

> A 128-bit cluster also improves our "resolution" for interoperability with other systems that want to use UUIDs for cluster ID to allow for random generation of cluster ID without the cost of having to introduce a centralized ID coordinator. Plus it gives us a further security guarantee that clusters can't talk to other clusters unless they can literally “guess” the identifier, which is now cryptographically "tough" (granted, we might not be able to imagine now why this could be important, but it's a stronger property nevertheless—especially from an online verification point of view).

+++ src/clients/dotnet/TigerBeetle/Client.cs
@@ -11,10 +11,10 @@ namespace TigerBeetle
     {
         private const int DEFAULT_CONCURRENCY_MAX = 32;
 
-        private readonly uint clusterID;
+        private readonly UInt128 clusterID;
         private readonly NativeClient nativeClient;
 
-        public Client(uint clusterID, string[] addresses, int concurrencyMax = DEFAULT_CONCURRENCY_MAX)
+        public Client(UInt128 clusterID, string[] addresses, int concurrencyMax = DEFAULT_CONCURRENCY_MAX)
         {
             this.nativeClient = NativeClient.Init(clusterID, addresses, concurrencyMax);
             this.clusterID = clusterID;
@@ -29,7 +29,7 @@ namespace TigerBeetle
             }
         }
 
-        public uint ClusterID => clusterID;
+        public UInt128 ClusterID => clusterID;
 
         public CreateAccountResult CreateAccount(Account account)
         {
